Phase 4d: extract NervesHub.Devices.Certificates sub-context
Final step of splitting the Devices god-module. Built on Phase 4c
(cleanup/phase-4c-devices-health).

Moves the 11 device-certificate functions out of NervesHub.Devices into a new,
documented NervesHub.Devices.Certificates module:
create_device_certificate/2, has_device_certificates?/1, get_device_certificates/1,
preload_device_certificates/2, get_device_by_certificate/1, get_device_by_x509/1,
get_device_certificate_by_x509/1, get_device_by_public_key/1,
get_device_certificate_by_device_and_serial/2, update_device_certificate/2,
delete_device_certificate/1.

The DeviceCertificate schema stays referenced by Devices (soft-delete cleanup
query), but the now-unused Certificate alias is dropped from Devices.

Function names unchanged; 37 call sites across 11 lib/test files repointed to
the new module.

With this, the Devices context's four sub-domains (CACertificates, Pinning,
Health, Certificates) are each in their own module.

Verification: mix format, credo (new module clean), dialyzer (0 unskipped),
full test suite (1396 passing).
