+++ lib/nerves_hub/devices.ex
@@ -11,7 +11,6 @@ defmodule NervesHub.Devices do
   alias NervesHub.Accounts.User
   alias NervesHub.AuditLogs
   alias NervesHub.AuditLogs.DeviceTemplates
-  alias NervesHub.Certificate
   alias NervesHub.DeploymentOrchestratorEvents
   alias NervesHub.DeviceEvents
   alias NervesHub.DeviceLink.DeviceInfo
@@ -278,13 +277,6 @@ defmodule NervesHub.Devices do
     |> Repo.one!()
   end
 
-  @doc """
-  Preloads a device's certificates. Pass `force: true` to reload them.
-  """
-  def preload_device_certificates(%Device{} = device, opts \\ []) do
-    Repo.preload(device, :device_certificates, opts)
-  end
-
   defp join_and_preload_deployment_group_and_current_release(query) do
     query
     |> join(:left, [d], dp in assoc(d, :deployment_group), as: :deployment_group)
@@ -329,17 +321,6 @@ defmodule NervesHub.Devices do
     query
     |> join(:left, [d], p in assoc(d, :product), as: :product)
     |> preload([product: p], product: p)
-  end
-
-  def get_device_by_x509(cert) do
-    fingerprint = NervesHub.Certificate.fingerprint(cert)
-
-    Device
-    |> join(:inner, [d], p in assoc(d, :product))
-    |> join(:inner, [d], dc in assoc(d, :device_certificates))
-    |> where([_, _, dc], dc.fingerprint == ^fingerprint)
-    |> preload([_d, p], product: p)
-    |> Repo.fetch()
   end
 
   @spec get_shared_secret_auth(String.t()) ::
@@ -420,113 +401,6 @@ defmodule NervesHub.Devices do
 
   def destroy_device(%Device{} = device) do
     Repo.delete(device)
-  end
-
-  @spec create_device_certificate(Device.t(), map() | X509.Certificate.t()) ::
-          {:ok, DeviceCertificate.t()}
-          | {:error, Changeset.t()}
-  def create_device_certificate(%Device{} = device, otp_cert) when is_tuple(otp_cert) do
-    {nb, na} = Certificate.get_validity(otp_cert)
-
-    params = %{
-      aki: Certificate.get_aki(otp_cert),
-      der: Certificate.to_der(otp_cert),
-      not_after: na,
-      not_before: nb,
-      serial: Certificate.get_serial_number(otp_cert),
-      ski: Certificate.get_ski(otp_cert)
-    }
-
-    create_device_certificate(device, params)
-  end
-
-  def create_device_certificate(%Device{} = device, params) do
-    params = Map.put(params, :org_id, device.org_id)
-
-    changeset =
-      device
-      |> Ecto.build_assoc(:device_certificates)
-      |> DeviceCertificate.changeset(params)
-
-    case Repo.insert(changeset) do
-      {:ok, device_certificate} ->
-        :telemetry.execute([:nerves_hub, :device_certificates, :created], %{count: 1})
-
-        {:ok, device_certificate}
-
-      {:error, error} ->
-        {:error, error}
-    end
-  end
-
-  def has_device_certificates?(%Device{} = device) do
-    DeviceCertificate
-    |> join(:inner, [dc], d in assoc(dc, :device))
-    |> where([_dc, d], d.id == ^device.id)
-    |> Repo.exists?()
-  end
-
-  def get_device_certificates(%Device{} = device) do
-    DeviceCertificate
-    |> join(:inner, [dc], d in assoc(dc, :device))
-    |> where([_dc, d], d.id == ^device.id)
-    |> Repo.all()
-  end
-
-  @spec get_device_by_certificate(DeviceCertificate.t()) ::
-          {:ok, Device.t()} | {:error, :not_found}
-  def get_device_by_certificate(%DeviceCertificate{device: %Ecto.Association.NotLoaded{}} = cert) do
-    Repo.preload(cert, :device)
-    |> get_device_by_certificate()
-  end
-
-  def get_device_by_certificate(%DeviceCertificate{device: %Device{} = device}), do: {:ok, Repo.preload(device, :org)}
-
-  def get_device_by_certificate(_), do: {:error, :not_found}
-
-  def get_device_certificate_by_x509(cert) do
-    fingerprint = NervesHub.Certificate.fingerprint(cert)
-
-    DeviceCertificate
-    |> where(fingerprint: ^fingerprint)
-    |> join(:inner, [dc], d in assoc(dc, :device))
-    |> preload([_dc, d], device: d)
-    |> Repo.fetch()
-  end
-
-  def get_device_by_public_key(otp_cert) do
-    pk_fingerprint = NervesHub.Certificate.public_key_fingerprint(otp_cert)
-
-    Device
-    |> join(:inner, [d], dc in assoc(d, :device_certificates))
-    |> where([_d, dc], dc.public_key_fingerprint == ^pk_fingerprint)
-    |> Repo.one()
-  end
-
-  @spec get_device_certificate_by_device_and_serial(Device.t(), binary) ::
-          {:ok, DeviceCertificate.t()} | {:error, any()}
-  def get_device_certificate_by_device_and_serial(%Device{id: device_id}, serial) do
-    query =
-      from(
-        dc in DeviceCertificate,
-        where: dc.serial == ^serial and dc.device_id == ^device_id
-      )
-
-    query
-    |> Repo.fetch()
-  end
-
-  def update_device_certificate(%DeviceCertificate{} = certificate, params) do
-    certificate
-    |> DeviceCertificate.update_changeset(params)
-    |> Repo.update()
-  end
-
-  @spec delete_device_certificate(DeviceCertificate.t()) ::
-          {:ok, DeviceCertificate.t()}
-          | {:error, Changeset.t()}
-  def delete_device_certificate(%DeviceCertificate{} = device_certificate) do
-    Repo.delete(device_certificate)
   end
 
   def clean_up_soft_deleted_devices() do
